replaced KL divergence with Wasserstein distance
The KL divergence has the issue that it is not symmetric, i.e., it penalises the
approximating distribution for not covering probability mass in the true
distribution, but more or less allows the approximating distribution to cover
regions of the space which are not covered by the true distribution. The
Wasserstein distance, or earth-mover's distance is a true, symmetric distance
between probability distributions and can even be computed more easily for
Gaussians than the KL divergence.

I have replaced all occurrences of the KL divergence with the Wasserstein
distance in code and text. Some of the results became clearer, for example,
because the scaled set is now penalised for overestimating variances. The
comparison for Gaussian vs. non-Gaussian source distributions now shows evidence
in favor of the hypothesis that capturing 4th moments helps in approximating the
transformed distribution, but the overall conclusions about good parameter
settings remain the same. The perhaps biggest change is that the UT appears
to be really quite inaccurate in high dimensions when measured with the
Wasserstein distance which leads to the counterintuitive situation that you only
need very few random samples to get better approximations of transformed mean
and covariance than the UT. This result questions the usefulness of the UT
and UKF in high dimensions.